Add AdminService and repository interfaces for hotel management
This commit introduces the `AdminService` class implementing the `IAdminService` interface, which includes methods for creating hotels, rooms, room types, and discounts.

The service interacts with the data layer through the `IAdminRepository`, which has been defined to check for existing cities and hotels, as well as to add new entities to the database.

The `Program.cs` file is updated to register the `IAdminRepository` and `AdminRepository` with the dependency injection container.

**NOTE: `AdminController` will be refactored later as it depends on most of the other services and their CRUD operations.